Move API to "experimental" and clean up Swagger leftovers
