+++ clients/java-legacy/api/openapi.yaml
@@ -3355,7 +3355,7 @@ paths:
           description: Internal Server Error
       summary: hard reset branch
       tags:
-      - branches
+      - experimental
       x-accepts: application/json
   /repositories/{repository}/branches/{branch}/revert:
     post:
@@ -7013,35 +7013,15 @@ components:
     ResetCreation:
       example:
         path: path
-        ref: ref
         force: false
         type: object
-        operation: staged
       properties:
-        operation:
-          default: staged
-          description: |
-            The kind of reset operation to perform.  If "staged",
-            uncommitted objects according to type.  If "hard", branch must
-            contain no uncommitted objects, and will be reset to refer to
-            ref.
-          enum:
-          - staged
-          - hard
-          type: string
         type:
-          description: |
-            Only allowed for operation="staged".  Specifies what to reset
-            according to path.
+          description: What to reset according to path.
           enum:
           - object
           - common_prefix
           - reset
-          type: string
-        ref:
-          description: |
-            Only allowed for operation="hard".  Branch will be reset to this
-            ref.
           type: string
         path:
           type: string
