+++ clients/java-legacy/src/test/java/io/lakefs/clients/api/BranchesApiTest.java
@@ -127,24 +127,6 @@ public class BranchesApiTest {
     }
     
     /**
-     * hard reset branch
-     *
-     * Relocate branch to refer to ref.  Branch must not contain uncommitted data.
-     *
-     * @throws ApiException
-     *          if the Api call fails
-     */
-    @Test
-    public void hardResetBranchTest() throws ApiException {
-        String repository = null;
-        String branch = null;
-        String ref = null;
-        Boolean force = null;
-                api.hardResetBranch(repository, branch, ref, force);
-        // TODO: test validations
-    }
-    
-    /**
      * list branches
      *
      *
