+++ clients/java-legacy/src/test/java/io/lakefs/clients/api/ExperimentalApiTest.java
@@ -110,6 +110,24 @@ public class ExperimentalApiTest {
     }
     
     /**
+     * hard reset branch
+     *
+     * Relocate branch to refer to ref.  Branch must not contain uncommitted data.
+     *
+     * @throws ApiException
+     *          if the Api call fails
+     */
+    @Test
+    public void hardResetBranchTest() throws ApiException {
+        String repository = null;
+        String branch = null;
+        String ref = null;
+        Boolean force = null;
+                api.hardResetBranch(repository, branch, ref, force);
+        // TODO: test validations
+    }
+    
+    /**
      * perform otf diff
      *
      *
