+++ clients/java/src/main/java/io/lakefs/clients/sdk/BranchesApi.java
@@ -1101,220 +1101,6 @@ public class BranchesApi {
     public APIgetBranchRequest getBranch(String repository, String branch) {
         return new APIgetBranchRequest(repository, branch);
     }
-    private okhttp3.Call hardResetBranchCall(String repository, String branch, String ref, Boolean force, final ApiCallback _callback) throws ApiException {
-        String basePath = null;
-        // Operation Servers
-        String[] localBasePaths = new String[] {  };
-
-        // Determine Base Path to Use
-        if (localCustomBaseUrl != null){
-            basePath = localCustomBaseUrl;
-        } else if ( localBasePaths.length > 0 ) {
-            basePath = localBasePaths[localHostIndex];
-        } else {
-            basePath = null;
-        }
-
-        Object localVarPostBody = null;
-
-        // create path and map variables
-        String localVarPath = "/repositories/{repository}/branches/{branch}/hard_reset"
-            .replace("{" + "repository" + "}", localVarApiClient.escapeString(repository.toString()))
-            .replace("{" + "branch" + "}", localVarApiClient.escapeString(branch.toString()));
-
-        List<Pair> localVarQueryParams = new ArrayList<Pair>();
-        List<Pair> localVarCollectionQueryParams = new ArrayList<Pair>();
-        Map<String, String> localVarHeaderParams = new HashMap<String, String>();
-        Map<String, String> localVarCookieParams = new HashMap<String, String>();
-        Map<String, Object> localVarFormParams = new HashMap<String, Object>();
-
-        if (ref != null) {
-            localVarQueryParams.addAll(localVarApiClient.parameterToPair("ref", ref));
-        }
-
-        if (force != null) {
-            localVarQueryParams.addAll(localVarApiClient.parameterToPair("force", force));
-        }
-
-        final String[] localVarAccepts = {
-            "application/json"
-        };
-        final String localVarAccept = localVarApiClient.selectHeaderAccept(localVarAccepts);
-        if (localVarAccept != null) {
-            localVarHeaderParams.put("Accept", localVarAccept);
-        }
-
-        final String[] localVarContentTypes = {
-        };
-        final String localVarContentType = localVarApiClient.selectHeaderContentType(localVarContentTypes);
-        if (localVarContentType != null) {
-            localVarHeaderParams.put("Content-Type", localVarContentType);
-        }
-
-        String[] localVarAuthNames = new String[] { "basic_auth", "cookie_auth", "oidc_auth", "saml_auth", "jwt_token" };
-        return localVarApiClient.buildCall(basePath, localVarPath, "PUT", localVarQueryParams, localVarCollectionQueryParams, localVarPostBody, localVarHeaderParams, localVarCookieParams, localVarFormParams, localVarAuthNames, _callback);
-    }
-
-    @SuppressWarnings("rawtypes")
-    private okhttp3.Call hardResetBranchValidateBeforeCall(String repository, String branch, String ref, Boolean force, final ApiCallback _callback) throws ApiException {
-        // verify the required parameter 'repository' is set
-        if (repository == null) {
-            throw new ApiException("Missing the required parameter 'repository' when calling hardResetBranch(Async)");
-        }
-
-        // verify the required parameter 'branch' is set
-        if (branch == null) {
-            throw new ApiException("Missing the required parameter 'branch' when calling hardResetBranch(Async)");
-        }
-
-        // verify the required parameter 'ref' is set
-        if (ref == null) {
-            throw new ApiException("Missing the required parameter 'ref' when calling hardResetBranch(Async)");
-        }
-
-        return hardResetBranchCall(repository, branch, ref, force, _callback);
-
-    }
-
-
-    private ApiResponse<Void> hardResetBranchWithHttpInfo(String repository, String branch, String ref, Boolean force) throws ApiException {
-        okhttp3.Call localVarCall = hardResetBranchValidateBeforeCall(repository, branch, ref, force, null);
-        return localVarApiClient.execute(localVarCall);
-    }
-
-    private okhttp3.Call hardResetBranchAsync(String repository, String branch, String ref, Boolean force, final ApiCallback<Void> _callback) throws ApiException {
-
-        okhttp3.Call localVarCall = hardResetBranchValidateBeforeCall(repository, branch, ref, force, _callback);
-        localVarApiClient.executeAsync(localVarCall, _callback);
-        return localVarCall;
-    }
-
-    public class APIhardResetBranchRequest {
-        private final String repository;
-        private final String branch;
-        private final String ref;
-        private Boolean force;
-
-        private APIhardResetBranchRequest(String repository, String branch, String ref) {
-            this.repository = repository;
-            this.branch = branch;
-            this.ref = ref;
-        }
-
-        /**
-         * Set force
-         * @param force  (optional, default to false)
-         * @return APIhardResetBranchRequest
-         */
-        public APIhardResetBranchRequest force(Boolean force) {
-            this.force = force;
-            return this;
-        }
-
-        /**
-         * Build call for hardResetBranch
-         * @param _callback ApiCallback API callback
-         * @return Call to execute
-         * @throws ApiException If fail to serialize the request body object
-         * @http.response.details
-         <table summary="Response Details" border="1">
-            <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-            <tr><td> 204 </td><td> reset successful </td><td>  -  </td></tr>
-            <tr><td> 400 </td><td> Bad Request </td><td>  -  </td></tr>
-            <tr><td> 401 </td><td> Unauthorized </td><td>  -  </td></tr>
-            <tr><td> 403 </td><td> Forbidden </td><td>  -  </td></tr>
-            <tr><td> 404 </td><td> Resource Not Found </td><td>  -  </td></tr>
-            <tr><td> 420 </td><td> too many requests </td><td>  -  </td></tr>
-            <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
-         </table>
-         */
-        public okhttp3.Call buildCall(final ApiCallback _callback) throws ApiException {
-            return hardResetBranchCall(repository, branch, ref, force, _callback);
-        }
-
-        /**
-         * Execute hardResetBranch request
-         * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
-         * @http.response.details
-         <table summary="Response Details" border="1">
-            <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-            <tr><td> 204 </td><td> reset successful </td><td>  -  </td></tr>
-            <tr><td> 400 </td><td> Bad Request </td><td>  -  </td></tr>
-            <tr><td> 401 </td><td> Unauthorized </td><td>  -  </td></tr>
-            <tr><td> 403 </td><td> Forbidden </td><td>  -  </td></tr>
-            <tr><td> 404 </td><td> Resource Not Found </td><td>  -  </td></tr>
-            <tr><td> 420 </td><td> too many requests </td><td>  -  </td></tr>
-            <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
-         </table>
-         */
-        public void execute() throws ApiException {
-            hardResetBranchWithHttpInfo(repository, branch, ref, force);
-        }
-
-        /**
-         * Execute hardResetBranch request with HTTP info returned
-         * @return ApiResponse&lt;Void&gt;
-         * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
-         * @http.response.details
-         <table summary="Response Details" border="1">
-            <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-            <tr><td> 204 </td><td> reset successful </td><td>  -  </td></tr>
-            <tr><td> 400 </td><td> Bad Request </td><td>  -  </td></tr>
-            <tr><td> 401 </td><td> Unauthorized </td><td>  -  </td></tr>
-            <tr><td> 403 </td><td> Forbidden </td><td>  -  </td></tr>
-            <tr><td> 404 </td><td> Resource Not Found </td><td>  -  </td></tr>
-            <tr><td> 420 </td><td> too many requests </td><td>  -  </td></tr>
-            <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
-         </table>
-         */
-        public ApiResponse<Void> executeWithHttpInfo() throws ApiException {
-            return hardResetBranchWithHttpInfo(repository, branch, ref, force);
-        }
-
-        /**
-         * Execute hardResetBranch request (asynchronously)
-         * @param _callback The callback to be executed when the API call finishes
-         * @return The request call
-         * @throws ApiException If fail to process the API call, e.g. serializing the request body object
-         * @http.response.details
-         <table summary="Response Details" border="1">
-            <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-            <tr><td> 204 </td><td> reset successful </td><td>  -  </td></tr>
-            <tr><td> 400 </td><td> Bad Request </td><td>  -  </td></tr>
-            <tr><td> 401 </td><td> Unauthorized </td><td>  -  </td></tr>
-            <tr><td> 403 </td><td> Forbidden </td><td>  -  </td></tr>
-            <tr><td> 404 </td><td> Resource Not Found </td><td>  -  </td></tr>
-            <tr><td> 420 </td><td> too many requests </td><td>  -  </td></tr>
-            <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
-         </table>
-         */
-        public okhttp3.Call executeAsync(final ApiCallback<Void> _callback) throws ApiException {
-            return hardResetBranchAsync(repository, branch, ref, force, _callback);
-        }
-    }
-
-    /**
-     * hard reset branch
-     * Relocate branch to refer to ref.  Branch must not contain uncommitted data.
-     * @param repository  (required)
-     * @param branch  (required)
-     * @param ref After reset, branch will point at this reference. (required)
-     * @return APIhardResetBranchRequest
-     * @http.response.details
-     <table summary="Response Details" border="1">
-        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-        <tr><td> 204 </td><td> reset successful </td><td>  -  </td></tr>
-        <tr><td> 400 </td><td> Bad Request </td><td>  -  </td></tr>
-        <tr><td> 401 </td><td> Unauthorized </td><td>  -  </td></tr>
-        <tr><td> 403 </td><td> Forbidden </td><td>  -  </td></tr>
-        <tr><td> 404 </td><td> Resource Not Found </td><td>  -  </td></tr>
-        <tr><td> 420 </td><td> too many requests </td><td>  -  </td></tr>
-        <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
-     </table>
-     */
-    public APIhardResetBranchRequest hardResetBranch(String repository, String branch, String ref) {
-        return new APIhardResetBranchRequest(repository, branch, ref);
-    }
     private okhttp3.Call listBranchesCall(String repository, String prefix, String after, Integer amount, final ApiCallback _callback) throws ApiException {
         String basePath = null;
         // Operation Servers
