+++ clients/java/src/test/java/io/lakefs/clients/sdk/model/ResetCreationTest.java
@@ -38,27 +38,11 @@ public class ResetCreationTest {
     }
 
     /**
-     * Test the property 'operation'
-     */
-    @Test
-    public void operationTest() {
-        // TODO: test operation
-    }
-
-    /**
      * Test the property 'type'
      */
     @Test
     public void typeTest() {
         // TODO: test type
-    }
-
-    /**
-     * Test the property 'ref'
-     */
-    @Test
-    public void refTest() {
-        // TODO: test ref
     }
 
     /**
