+++ clients/python-legacy/test/test_branches_api.py
@@ -59,13 +59,6 @@ class TestBranchesApi(unittest.TestCase):
         """
         pass
 
-    def test_hard_reset_branch(self):
-        """Test case for hard_reset_branch
-
-        hard reset branch  # noqa: E501
-        """
-        pass
-
     def test_list_branches(self):
         """Test case for list_branches
