+++ clients/python/test/test_branches_api.py
@@ -64,13 +64,6 @@ class TestBranchesApi(unittest.TestCase):
         """
         pass
 
-    def test_hard_reset_branch(self):
-        """Test case for hard_reset_branch
-
-        hard reset branch  # noqa: E501
-        """
-        pass
-
     def test_list_branches(self):
         """Test case for list_branches
